CORS/tests/submissions/Microsoft/CORS_009.htm

<!DOCTYPE html>
<html>
    <head>
        <title>XHR Requests fail if &quot;Access-Control-Allow-Origin: *&quot; and withCredentials is true</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta description="XHR Requests fail if &quot;Access-Control-Allow-Origin: *&quot; and withCredentials is true" />
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
        <link rel="help" href="http://dvcs.w3.org/hg/cors/raw-file/tip/Overview.html#resource-requests"/>
        <script src="http://w3c-test.org/resources/testharness.js"></script>
        <script src="http://w3c-test.org/resources/testharnessreport.js"></script>
        <script src="SetCORSUrls.js"></script>
    </head>
    <body>
        <div id=log></div>
        <script type="text/javascript">
            var t = async_test("XHR Requests fail if 'Access-Control-Allow-Origin: *' and withCredentials is true");
            var xhr = new XMLHttpRequest();
            if('withCredentials' in xhr)
            {
                var url = CORS_SERVER_URL + "?TestID=9"
                xhr.open("GET", url, true);
                xhr.withCredentials = true;
                xhr.onreadystatechange = function(e)
                {
                    if (xhr.readyState == 4)
                    {
                        // Server responds back with the header 'Access-Control-Allow-Origin: *'
                        // The request should fail because the string "*" cannot be used for a
                        // resource that supports credentials.
                        t.step(function(){assert_true(xhr.status === 0 && xhr.responseText === "")});
                        t.done();
                    }
                }
                xhr.send();
            }
            else
            {
                t.step(function(){assert_unreached("[Cross-Origin XMLHttpRequests are not supported]")});
                t.done();
            }
        </script>
    </body>
</html>
